Separate test builders from top level logic
IMO it's easier to follow when:
- the tests are separated from the test setup
- there's fewer nested scopes
